Performance overhaul of the Encoding API.
This comprises a number of changes to how serialization
is handled:

1. All of the encoders are now structs that are passed
   as `inout` parameters.

2. `traverse` methods are now generic in the visitor type
   instead of taking a protocol type.  This eliminates
   the creation of protocol witnesses.

3. `traverse` now calls type-specific methods directly,
   e.g., `visitSingularStringField()`.  This significantly
   shortens the call depth and eliminates generic reflection.
   This is a big overall speedup for binary coding.

4. Text and JSON now construct UTF8 byte sequences directly,
   instead of appending to strings.

5. Text and JSON now use `StaticString` field names.

6. Many small micro-optimizations throughout.

Here are a few numbers measured on my laptop:

"100 int64":      C++  Master  This PR
Encode Binary   1.667   8.343    3.772
Encode JSON    78.684 168.201   28.859
Encode Text    17.912 104.672   26.875

"100 string":     C++  Master  This PR
Encode Binary   5.350  58.713   13.625
Encode JSON    77.743 239.404   29.504
Encode Text    18.358 174.878   29.545

"100 double":     C++  Master  This PR
Encode Binary   0.921   8.099    3.748
Encode JSON    95.823 128.070   24.358
Encode Text    38.967 100.073   25.291

(This last is cheating a little; our JSON/Text encoders now use a fast
path if the double value happens to be an integer, and our test suite
only uses integral values for exercising numeric fields.)

Bonus: Most of the encoder-specific `FieldType` methods are gone,
replaced by a standard set of abstract reflectors used for maps and
extensions.

Additional Bonus: All five encoding visitors and all but one of the
low-level encoders are now `internal` to the library.  The only
exception is `JSONEncoder` which is public but has no public members.

+++ Sources/SwiftProtobuf/Google_Protobuf_Timestamp_Extensions.swift
@@ -133,7 +133,7 @@ private func parseTimestamp(s: String) throws -> (Int64, Int32) {
         pos += 1
         var digitValue = 100000000
         while pos < value.count && value[pos] >= zero && value[pos] <= nine {
-            nanos += digitValue * (value[pos] - zero)
+            nanos += Int32(digitValue * (value[pos] - zero))
             digitValue /= 10
             pos += 1
         }
